+++ app/src/main/java/com/example/modules/dashboard/logic/DebtLedgerCalculator.kt
@@ -8,18 +8,15 @@ import kotlin.math.abs
 data class WalletDebtLedgerItem(
     val wallet: WalletAccount,
     val owner: Member?,
-    val totalTransferIn: Double,
-    val totalTransferOut: Double,
-    val totalExpensePaid: Double,
-    val netDebtBalance: Double // (Transfer In) - (Transfer Out) - (Expense Paid)
+    val totalTransferIn: Long,
+    val totalTransferOut: Long,
+    val totalExpensePaid: Long,
+    val netDebtBalance: Long // (Transfer In) - (Transfer Out) - (Expense Paid)
 ) {
-    // Positif: Partner masih pegang uangmu
-    // Negatif: Pengeluaran melebihi transfer, kamu berhutang ke partner
-    // Nol: Lunas / Imbang
     val statusText: String
         get() = when {
-            netDebtBalance > 0 -> "Partner Pegang Uangmu"
-            netDebtBalance < 0 -> "Hutang ke Partner (Reimburse)"
+            netDebtBalance > 0L -> "Partner Pegang Uangmu"
+            netDebtBalance < 0L -> "Hutang ke Partner (Reimburse)"
             else -> "Lunas / Terbayar"
         }
 }
@@ -34,17 +31,14 @@ object DebtLedgerCalculator {
             val owner = members.find { it.id == wallet.memberId }
             val walletTxs = transactions.filter { it.walletId == wallet.id }
 
-            // Transfer In: Masuk ke wallet ini dari transfer
             val transferIn = walletTxs
                 .filter { it.amount > 0 && (it.note.contains("transfer", ignoreCase = true) || it.categoryId.contains("tf", ignoreCase = true)) }
                 .sumOf { it.amount }
 
-            // Transfer Out: Keluar dari wallet ini untuk transfer
             val transferOut = walletTxs
                 .filter { it.amount < 0 && (it.note.contains("transfer", ignoreCase = true) || it.categoryId.contains("tf", ignoreCase = true)) }
                 .sumOf { abs(it.amount) }
 
-            // Expense Paid: Pengeluaran ril yang dibayar menggunakan wallet ini
             val expensePaid = walletTxs
                 .filter { it.amount < 0 && !it.note.contains("transfer", ignoreCase = true) && !it.categoryId.contains("tf", ignoreCase = true) }
                 .sumOf { abs(it.amount) }
@@ -62,3 +56,4 @@ object DebtLedgerCalculator {
         }
     }
 }
+
