+++ app/src/main/java/com/example/modules/dashboard/logic/SavingsIntegrityCalculator.kt
@@ -6,17 +6,17 @@ import kotlin.math.abs
 
 data class SavingsCategoryDetail(
     val category: Category,
-    val totalSaved: Double,
-    val totalSpentDiverted: Double,
-    val netBalance: Double,
+    val totalSaved: Long,
+    val totalSpentDiverted: Long,
+    val netBalance: Long,
     val integrityRate: Double,
     val isCompromised: Boolean
 )
 
 data class SavingsIntegrityReport(
-    val totalSavingsAllocated: Double,
-    val totalSavingsUsed: Double,
-    val netSavingsRetained: Double,
+    val totalSavingsAllocated: Long,
+    val totalSavingsUsed: Long,
+    val netSavingsRetained: Long,
     val overallIntegrityRate: Double,
     val compromisedCount: Int,
     val categoryDetails: List<SavingsCategoryDetail>
@@ -34,19 +34,17 @@ object SavingsIntegrityCalculator {
         val details = savingsCategories.map { cat ->
             val catTxs = transactions.filter { it.categoryId == cat.id }
             
-            // Inflow: Positive transactions or expenses intended as savings allocation
             val positiveTxs = catTxs.filter { it.amount > 0 }.sumOf { it.amount }
             val negativeTxs = catTxs.filter { it.amount < 0 }.sumOf { abs(it.amount) }
             
-            // In the CSV structure, savings entries recorded as expense with note "Save#1" / "Komitmen" are saved money
-            val totalSaved = if (positiveTxs > 0) positiveTxs else negativeTxs
-            val totalSpentDiverted = if (positiveTxs > 0) negativeTxs else catTxs.filter { 
+            val totalSaved = if (positiveTxs > 0L) positiveTxs else negativeTxs
+            val totalSpentDiverted = if (positiveTxs > 0L) negativeTxs else catTxs.filter { 
                 it.amount < 0 && (it.note.contains("pakai", ignoreCase = true) || it.note.contains("tarik", ignoreCase = true) || it.note.contains("bocor", ignoreCase = true))
             }.sumOf { abs(it.amount) }
 
             val netBalance = totalSaved - totalSpentDiverted
-            val rate = if (totalSaved > 0) ((netBalance / totalSaved) * 100.0).coerceIn(0.0, 100.0) else 100.0
-            val compromised = totalSpentDiverted > 0
+            val rate = if (totalSaved > 0L) ((netBalance.toDouble() / totalSaved.toDouble()) * 100.0).coerceIn(0.0, 100.0) else 100.0
+            val compromised = totalSpentDiverted > 0L
 
             SavingsCategoryDetail(
                 category = cat,
@@ -61,7 +59,7 @@ object SavingsIntegrityCalculator {
         val totalAllocated = details.sumOf { it.totalSaved }
         val totalUsed = details.sumOf { it.totalSpentDiverted }
         val netRetained = totalAllocated - totalUsed
-        val overallRate = if (totalAllocated > 0) ((netRetained / totalAllocated) * 100.0).coerceIn(0.0, 100.0) else 100.0
+        val overallRate = if (totalAllocated > 0L) ((netRetained.toDouble() / totalAllocated.toDouble()) * 100.0).coerceIn(0.0, 100.0) else 100.0
         val compromisedCount = details.count { it.isCompromised }
 
         return SavingsIntegrityReport(
@@ -74,3 +72,4 @@ object SavingsIntegrityCalculator {
         )
     }
 }
+
